requirements.txt
streamlit
python-dotenv
pypdf
PyPDF2
langchain
langchain-groq
langchain-pinecone
langchain-community
pinecone-client
faiss-cpu
sentence-transformers
